+++ src/SepomexPhp/District.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace SepomexPhp;
+namespace Eclipxe\SepomexPhp;
 
-use SepomexPhp\Traits\PropertyIdIntegerTrait;
-use SepomexPhp\Traits\PropertyNameStringTrait;
-use SepomexPhp\Traits\PropertyStateTrait;
+use Eclipxe\SepomexPhp\Traits\PropertyIdIntegerTrait;
+use Eclipxe\SepomexPhp\Traits\PropertyNameStringTrait;
+use Eclipxe\SepomexPhp\Traits\PropertyStateTrait;
 
 class District
 {