+++ src/SepomexPhp/Location.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace SepomexPhp;
+namespace Eclipxe\SepomexPhp;
 
-use SepomexPhp\Traits\PropertyCityTrait;
-use SepomexPhp\Traits\PropertyDistrictTrait;
-use SepomexPhp\Traits\PropertyIdIntegerTrait;
-use SepomexPhp\Traits\PropertyNameStringTrait;
+use Eclipxe\SepomexPhp\Traits\PropertyCityTrait;
+use Eclipxe\SepomexPhp\Traits\PropertyDistrictTrait;
+use Eclipxe\SepomexPhp\Traits\PropertyIdIntegerTrait;
+use Eclipxe\SepomexPhp\Traits\PropertyNameStringTrait;
 
 class Location
 {